Ignore the stop command if the pid is missing

+++ lib/guard/unicorn.rb
@@ -86,6 +86,8 @@ module Guard
     end
 
     def stop_unicorn
+      return unless pid
+
       Process.kill("QUIT", pid) if Process.getpgid(pid) 
     end
